srd: Add experimental uart_dump decoder.
This PD stacks on top of the UART protocol decoder, and ouputs the
decoded UART data in the "canonical" format, one byte after the other
without any metadata (start bits, stop bits, parity bits, errors) and
without any sample-numbers and such. This is basically what you would
see in a terminal program, such as minicom.

Per default it outputs to stdout, but using the 'filename' option
and/or the 'filename_rx'/'filename_tx' options it can also output to
files.

Using the 'tx' and 'rx' options you can also select whether both, or
only one of the data direction's data should be output.

Example:

  sigrok-cli -i foo.sr \
             -a uart:rx=0:tx=1,uart_dump:rx=no:tx=yes:filename=foo.txt \
             -s uart,uart_dump

This PD is just an experiment for now, it's possible that it will be
replaced / reimplemented via a different sigrok facility later on.