srd: usb: Preliminary USB low-speed support.
Some of the differences of USB low-speed (compared to full-speed) are:

 - 1.5Mb/s datarate, not 12MBit/s.
 - The J and K states are swapped/inverted.

Note: This is work in progress, further changes may still be needed.

Add a 'signalling' option, which allows the user to select whether the
PD should decode using the low-speed or full-speed signalling protocol.
If unspecified, the default is full-speed.