uart: Fix corner-case that can occur with LA triggers.
Assume that the initial pin state is 1/high for the RX and TX lines.

This fixes the decode when an LA triggers on e.g. TX=low (the first
sample would be low in that case, so the falling edge for the start bit
would be missed by the decoder).