uart: Use same length for data output in all cases.
Until now you could get e.g. "d" or "df" as hex output from the UART PD.

This will now be a common-length "0D" and "DF". When all data byte
annotations are of the same lengths the readability in GUI traces is a lot
better. Also, hardcode hex characters to be upper-case (for now).

The same applies to oct ("021" instead of "21") and bin output
("00001001" instead of "1001").