uart: Output data in separate RX and TX annotation types.
There is one annotation type for the RX side of the communication/data
output now, and another one for the TX side.

This allows GUIs to show them in two different traces, and/or show them in
the same trace but with different colors.

This also has the additional benefit that it is clear which databyte
belongs to which direction. Previously the annotations had to look like
"RX: 5F" or "TX: 11001011", but the "RX: " and "TX: " prefixes are now
no longer required, making the GUI traces (and CLI output too) much more
readable.