spi: Make CS# optional and use it if supplied.
The CS# pin is now optional, it can either be supplied to the PD or not.

If it _is_ supplied, reset the PD state every time it changes. This has
the effect that "incomplete" SPI frames and those where CS# is not
asserted are ignored and not decoded.

This fixes bug #152.