Implement OUTPUT_META
This replaces the Decoder.add() method with Decoder.register().
The first argument is still output type, but all arguments are now
optional:

  Decoder.register(output_type,
        id='someid',
        meta=(object-type, 'Name', 'Description'))

'id' defaults to the protocol decoder instance id, and only needs changing
if a decoder chain needs to fork.
'object-type' refers to a Python object, such as int or str.

After registering, the PD submits data as usual with Decoder.put(), with
the only argument a value of the registered object-type.