Automate protocol decoder installation.
This automatically figures out the files to install for each protocol
decoder, without involving autotools.

All python files (filenames ending in .py) are always installed. If a
protocol decoder requires installation of a non-python file, a small
file called 'config' can be created in that protocol decoder's
directory, with the following content:

	# comments are ok
	extra-install vendorlist.txt commands.txt