Add protocol decoder testing framework.
This adds a tool in the tests directory, called pdtest. It uses the
"test/" directory in every PD directory, if present, to run the
PD against dumps found in the sigrok-dumps repository, and compares
the output against ".output" files in the "test/" directory. The file
"test/test.conf" is used to configure which tests to run.

A separate tool (tests/runtc.c) is used to run the actual decoding and
report output.

To get an overview of the options, run tests/pdtest without any options.