Add python as a pkg-config dependency
The current libsigrokdecode build passes the python
CPPFLAGS and LDFLAGS directly into libsigrokdecode.pc.in.
This leads to issues while cross-compiling as while using
PKG_CONFIG_SYSROOT_DIR, there is a chance of the prefix
being duplicated. We now, while detecting the installation
of python (cross or normal), add a dependency for the
module name in the pkgconfig file (from configure.ac).