runtc: Fix incorrect include directive
runtc.c used to try to include "../config.h".  This only works when
building in the source tree, otherwise there is no config.h in the
parent directory of runtc.c.  However, the explicit path is
unnecessary since the build directory (whether the same as the source
directory or not) will always be added to the include path.