srd_decoder_load_all_path: Support zipimport paths
srd_decoder_load() works perfectly with decoders stored inside zip files.
However, srd_decoder_load_all{_path,}() did not, because it assumed that
normal directory listings can be used to enumerate decoders, which is not
the case for zipped decoders.  Fix this by providing a fallback based on
the zipimporter Python class.