configure.ac: Use AM_CFLAGS instead of CFLAGS.
CFLAGS, CPPFLAGS, LDFLAGS and so on are so-called "user variables" and
are meant for users of the package to use/override. The package itself
should not set or modify them in any way. Instead, the "shadow variables"
such as AM_CFLAGS, AM_CPPFLAGS, AM_LDFLAGS and so on, can be modified by
the package.