Pass unitsize per sample chunk.
Don't pass unitsize to srd_inst_channel_set_all(), have that only
set the channel map. Instead, srd_session_send() now has a parameter
for the unitsize which is passed with every new chunk to be decoded.

This is in preparation to fix issues with devices or files which
have a unitsize != 1 and where the "guessed" unitsize based on the
number of channels is not correct.

This also allows for (potential) future changes where every chunk can
indeed have a different unitsize.

This fixes (parts of) bug #352.