Add srd_inst_initial_pins_set_all() and support code.
This allows frontends to set the assumed initial pins (i.e., the assumed
state of the pins before the first sample of a capture) to user-specified
values.

The assumed initial pins can be either low, or high, or "use same value
as the first sample of the capture".

The special self.initial_pins decoder attribute is now removed.