usb_power_delivery: enforce check order for start-of-packet sequences
The list of a dictionary's keys need not reproduce in identical order
everywhere. Make sure to run all start-of-packet sequence checks in the
decoder implementation in a specific order on each machine, such that
annotations get emitted with identical content and in the same order for
each execution of the decoder.

This fixes the remaining part of bug #1090.