Add a trivial "counter" decoder
This decoder just counts the number of falling and/or rising edges. This
is especially useful for diagnosing protocols with a clock signal or a
fixed number of transitions per bit, e.g. pulse length coded.

It also provides a divider, which can be used to e.g. count the number
of words in I²C or SPI transfers.