sdcard_spi: add support for CMD24 including payload data
The previous implementation advanced from the start of CMD24 to IDLE.
This commit introduces support for the data phase and data response of
CMD24. Which results in improved usability of the decoder output, and
reduced probability of false positives (don't detect "commands" in the
payload data).

This commit addresses the missing CMD24 handling part of bug #1365.