onewire_network: add a 'Disable Test Mode' command for DS2408
From the DS2408 datasheet [1], page 38:

"Power-up timing
The DS2408 is sensitive to the power-on slew rate and can inadvertently
power up with a test mode feature enabled. When this occurs, the P0 port
does not respond to the Channel Access Write command.

For most reliable operation, it is recommended to disable the test mode
after every power-on reset using the Disable Test Mode sequence shown
below. The 64-bit ROM code must be transmitted in the same bit sequence
as with the Match ROM command, i.e., least significant bit first."

This commit is adding this missing command so now a full decoding
is possible:
Annotations from 1-Wire network layer:
8990980-8991220 1-Wire network layer: : Reset/presence: true
8994163-8994437 1-Wire network layer: : ROM command: 0x96 'DS2408: Disable Test Mode'
8995026-9008178 1-Wire network layer: : ROM: 0xa30000001246a929
9008785-9009060 1-Wire network layer: : Data: 0x3c
9012290-9012530 1-Wire network layer: : Reset/presence: true

And from the specific stacked DS2408 slave:
8990980-8991220 DS2408: : Reset/presence: true
8995026-9008178 DS2408: : ROM: 0xa30000001246a929 (family code 0x29)
9008785-9009060 DS2408: : Disable Test Mode (0x3c)
9012290-9012530 DS2408: : Reset/presence: true

Refs:
[1] https://datasheets.maximintegrated.com/en/ds/DS2408.pdf