Initial Amulet LCD ASCII PD
This is a protocol decoder for the 'ASCII' protocol used by
Amulet Technologies LCDs.

Currently some commands are not implemented yet. I also lack capture data
from a display that will use replies other than ACK and NACK.

Reads are untested as I have no suitable captures.

The PD copes with bus errors (there is an actual bug in the device I'm
reverse engineering) and most of the commands are implemented.

The unimplemented commands should generally consume the correct
number of bytes from the bus, the exception to this are the drawing
commands, because there are actually at least two revisions of them
with different payloads, that are really hard to detect in greedy
algorithm.