spiflash: Add "EEPROM" to descriptions.
There are various different names that these types of memories are being
referred to in the wild: SPI flash, flash chip, flash, flash EEPROM,
SPI EEPROM, serial flash, serial memory, flash memory, and various others.

In order to make UI decoder selection more useful to the user, we add
the "EEPROM" string to some of the decoder metadata fields, so the
decoder will (for example) show up in PulseView's list of decoders when
the user types "eeprom" to narrow down the listed decoders.