irmp: introduce variables for start of IR frame and current sample
Introduce variables in the IRMP core logic which track the current
sample number, and the position where the start of an IR frame got
detected. The variables are conditional (ANALYZE builds only).

Provided by: Rene Staffen

Local modification: Drop the initializer for the static variables.
They reside in .bss and need not occupy .data space.